fix(plan): fold forecast dashboard into overview (CHAOS-2084)
